✅ boost coverage by factoring out copy mutable if needed
