+++ packages/atom.io/src/internal/get-state/read-or-compute-value.ts
@@ -7,7 +7,7 @@ export const readOrComputeValue = <T>(
 	state: ReadableState<T>,
 ): T => {
 	if (target.valueMap.has(state.key)) {
-		return readCachedValue(state, target)
+		return readCachedValue(target, state)
 	}
 	switch (state.type) {
 		case `readonly_held_selector`:
@@ -17,31 +17,24 @@ export const readOrComputeValue = <T>(
 			target.logger.info(`🧮`, state.type, state.key, `computing value`)
 			return state.get()
 		case `atom`: {
-			const def = state.default
-			let defaultValue: T
-			if (def instanceof Function) {
-				defaultValue = def()
+			let def: T
+			if (state.default instanceof Function) {
+				def = state.default()
 			} else {
-				defaultValue = def
+				def = state.default
 			}
-			const cachedValue = cacheValue(
-				target,
-				state.key,
-				defaultValue,
-				state.subject,
-			)
+			const cachedValue = cacheValue(target, state.key, def, state.subject)
 			target.logger.info(
 				`💁`,
 				`atom`,
 				state.key,
 				`could not find cached value; using default`,
-				defaultValue,
+				def,
 			)
 			return cachedValue
 		}
 		case `mutable_atom`: {
-			const Ctor = state.class
-			const instance = new Ctor()
+			const instance = new state.class()
 			const cachedValue = cacheValue(target, state.key, instance, state.subject)
 			target.logger.info(
 				`💁`,
