+++ packages/atom.io/src/internal/set-state/set-atom.ts
@@ -1,14 +1,13 @@
 import type { KeyedStateUpdate } from "atom.io"
 
-import type { Atom, Store } from ".."
+import type { Atom, MutableAtom, Store } from ".."
 import { hasRole } from "../atom/has-role"
-import { cacheValue } from "../caching"
+import { cacheValue, readCachedValue } from "../caching"
 import { readOrComputeValue } from "../get-state/read-or-compute-value"
 import { isTransceiver, type Transceiver } from "../mutable"
 import { markDone } from "../operation"
 import { isChildStore } from "../transaction/is-root-store"
 import { become } from "./become"
-import { copyMutableIfNeeded } from "./copy-mutable-if-needed"
 import { emitUpdate } from "./emit-update"
 import { evictDownStream } from "./evict-downstream"
 
@@ -18,13 +17,7 @@ export const setAtom = <T>(
 	next: T | ((oldValue: T) => T),
 ): void => {
 	const oldValue = readOrComputeValue(target, atom)
-	let newValue = oldValue
-	if (atom.type === `mutable_atom` && isChildStore(target)) {
-		const { parent } = target
-		const copiedValue = copyMutableIfNeeded(target, atom, parent)
-		newValue = copiedValue
-	}
-	newValue = become(next)(newValue)
+	let newValue = become(next)(oldValue)
 	target.logger.info(`📝`, `atom`, atom.key, `set to`, newValue)
 	newValue = cacheValue(target, atom.key, newValue, atom.subject)
 	markDone(target, atom.key)
@@ -60,14 +53,11 @@ export const setAtom = <T>(
 		)
 	} else if (hasRole(atom, `tracker:signal`)) {
 		const mutableKey = atom.key.slice(1)
-		const mutableAtom = target.atoms.get(mutableKey) as Atom<any>
-		let transceiver: Transceiver<any, any> = target.valueMap.get(mutableKey)
-		if (mutableAtom.type === `mutable_atom` && isChildStore(target)) {
-			const { parent } = target
-			const copiedValue = copyMutableIfNeeded(target, mutableAtom, parent)
-			transceiver = copiedValue
-		}
+		const mutable = target.atoms.get(mutableKey) as MutableAtom<any>
+		const transceiver: Transceiver<any, any> = readCachedValue(target, mutable)
 		const accepted = transceiver.do(update.newValue) === null
-		if (accepted) evictDownStream(target, mutableAtom)
+		if (accepted === true) {
+			evictDownStream(target, mutable)
+		}
 	}
 }
